Major changes
Data utils now correctly formats datetime into a string, and has new method to ensure value in a controller. Modified some methods in theme to add color option to headertext and added a new optionsSelector. Standardized the data class more and added a new abstract class for options when any class will be choses from a list of options. Added a projectType. Added an options page which directs to a new screen of options when tapping the text box for an input with options. Updated project page to include use of controllers for all fields, and added save and update capabilities using SampleData class. Update project form to use new options selectors. Added save and update functionality to work items as well as better use of controllers for all input fields. Redesigned the clock to work independently from the form itself and added better focus control when navigating between input fields.